forwprop: Also dce from added statements from gimple_simplify
This extends r14-3982-g9ea74d235c7e78 to also include the newly added statements
since some of them might be dead too (due to the way match and simplify works).
This was noticed while working on adding a new match and simplify pattern where a
new statement that got added was not being used.

Bootstrapped and tested on x86_64-linux-gnu with no regressions.

gcc/ChangeLog:

	* gimple-fold.cc (mark_lhs_in_seq_for_dce): New function.
	(replace_stmt_with_simplification): Call mark_lhs_in_seq_for_dce
	right before inserting the sequence.
	(fold_stmt_1): Add dce_worklist argument, update call to
	replace_stmt_with_simplification.
	(fold_stmt): Add dce_worklist argument, update call to fold_stmt_1.
	(fold_stmt_inplace): Update call to fold_stmt_1.
	* gimple-fold.h (fold_stmt): Add bitmap argument.
	* tree-ssa-forwprop.cc (pass_forwprop::execute): Update call to fold_stmt.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>