c++: Appertain standard attributes after array closing square bracket to array type rather than declarator [PR110345]
For C++ 26 P2552R3 I went through all the spots (except modules) where
attribute-specifier-seq appears in the grammar and tried to construct
a testcase in all those spots, for now for [[deprecated]] attribute.

This is the second issue I found.  The comment already correctly says that
attributes after closing ] appertain to the array type, but we were
appending them to returned_attrs, so effectively applying them to the
declarator (as if they appeared right after declarator-id).

2024-08-20  Jakub Jelinek  <jakub@redhat.com>

	PR c++/110345
	* decl.cc (grokdeclarator): Apply declarator->std_attributes
	for cdk_array to type, rather than chaining it to returned_attrs.

	* g++.dg/cpp0x/gen-attrs-82.C: New test.
	* g++.dg/gomp/attrs-3.C (foo): Expect different diagnostics for
	omp::directive attribute after closing square bracket of an automatic
	declaration and add a test with the attribute after array's
	declarator-id.