Remove unnecessary view_convert obsoleted by [PR86468].
This patch removes an unnecessary view_convert in trans_associate to
prevent hard to find runtime errors in the future.  The view_convert was
erroneously introduced not understanding why ranks of the arrays to
assign are different.  The ranks are fixed by PR86468 now and the
view_convert is obsolete.

gcc/fortran/ChangeLog:

	PR fortran/86468

	* trans-stmt.cc (trans_associate_var): Remove superfluous
	view_convert.