arm: [MVE intrinsics] fix checks of immediate arguments
As discussed in [1], it is better to use "su64" for immediates in
intrinsics signatures in order to provide better diagnostics
(erroneous constants are not truncated for instance).  This patch thus
uses su64 instead of ss32 in binary_lshift_unsigned,
binary_rshift_narrow, binary_rshift_narrow_unsigned, ternary_lshift,
ternary_rshift.

In addition, we fix cases where we called require_integer_immediate
whereas we just want to check that the argument is a scalar, and thus
use require_scalar_type in binary_acca_int32, binary_acca_int64,
unary_int32_acc.

Finally, in binary_lshift_unsigned we just want to check that 'imm' is
an immediate, not the optional predicates.

[1] https://gcc.gnu.org/pipermail/gcc-patches/2024-August/660262.html

2024-08-21  Christophe Lyon  <christophe.lyon@linaro.org>

	gcc/
	* config/arm/arm-mve-builtins-shapes.cc (binary_acca_int32): Fix
	check of scalar argument.
	(binary_acca_int64): Likewise.
	(binary_lshift_unsigned): Likewise.
	(binary_rshift_narrow): Likewise.
	(binary_rshift_narrow_unsigned): Likewise.
	(ternary_lshift): Likewise.
	(ternary_rshift): Likewise.
	(unary_int32_acc): Likewise.