Dump aliases in -fcallgraph-info
Dump ICF-unified decls, thunks, aliases and whatnot along with their
ultimate targets, with edges from the alias to the target.

Add support for dropping the source file's suffix when forming from
dump-base, so that auxiliary files can be scanned, such as the .ci
files generated by -fcallgraph-info, as in the testcase.


for  gcc/ChangeLog

	* toplev.cc (dump_final_alias_vcg): New.
	(dump_final_node_vcg): Dump aliases along with node.

for  gcc/testsuite/ChangeLog

	* lib/scandump.exp (dump-base): Support {} in dump base suffix
	to drop it.
	* gcc.dg/callgraph-info-1.c: New.