arm: [MVE intrinsics] factorize vdwdup viwdup
Factorize vdwdup and viwdup so that they use the same parameterized
names.

Like with vddup and vidup, we do not bother with the corresponding
expanders, as we stop using them in a subsequent patch.

The patch also adds the missing attributes to vdwdupq_wb_u_insn and
viwdupq_wb_u_insn patterns.

2024-08-21  Christophe Lyon  <christophe.lyon@linaro.org>

	gcc/
	* config/arm/iterators.md (mve_insn): Add VIWDUPQ, VDWDUPQ,
	VIWDUPQ_M, VDWDUPQ_M.
	(VIDWDUPQ): New iterator.
	(VIDWDUPQ_M): New iterator.
	* config/arm/mve.md (mve_vdwdupq_wb_u<mode>_insn)
	(mve_viwdupq_wb_u<mode>_insn): Merge into ...
	(@mve_<mve_insn>q_wb_u<mode>_insn): ... this. Add missing
	mve_unpredicated_insn and mve_move attributes.
	(mve_vdwdupq_m_wb_u<mode>_insn, mve_viwdupq_m_wb_u<mode>_insn):
	Merge into ...
	(@mve_<mve_insn>q_m_wb_u<mode>_insn): ... this.