testsuite: Fix gcc.dg/torture/pr116420.c for targets default unsigned char [PR116464]
This is an obvious fix to the gcc.dg/torture/pr116420.c testcase which simplier
changes from plain `char` to `signed char` so it works on targets where plain char defaults
to unsigned.

Pushed as obvious after a quick test for aarch64-linux-gnu to make sure the testcase
passes now.

	PR testsuite/116464

gcc/testsuite/ChangeLog:

	* gcc.dg/torture/pr116420.c:

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>