testsuite: Fix vect-mod-var.c for division by 0 [PR116461]
The testcase cc.dg/vect/vect-mod-var.c has an division by 0
which is undefined. On some targets (aarch64), the scalar and
the vectorized version, the result of division by 0 is the same.
While on other targets (x86), we get a SIGFAULT. On other targets (powerpc),
the results are different.

The fix is to make sure the testcase does not test division by 0 (or really mod by 0).

Pushed as obvious after testing on x86_64-linux-gnu to make sure the testcase passes
now.

	PR testsuite/116461

gcc/testsuite/ChangeLog:

	* gcc.dg/vect/vect-mod-var.c: Change the initialization loop so that
	`b[i]` is never 0. Use 1 in those places.

Signed-off-by: Andrew Pinski <quic_apinski@quicinc.com>